Changed nelements to int64_t. Made the loop variables same across all three correlation functions. Added some colored output to examples code. Python bindings is still not functional

+++ xi_theory/include/cellarray.h
@@ -2,19 +2,17 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 
 #include "function_precision.h" //definition of DOUBLE depending on Makefile flag -DOUBLE_PREC
 
 #define NLATMAX   100      /* maximum grid dimension in X-Y plane */
 
-
-//pointer alignment is 8 bytes. So the 4 bytes will be padded and wasted anyway ->
-//changing nelements to a int64_t is fine 
 typedef struct{
   DOUBLE *x;
   DOUBLE *y;
   DOUBLE *z;
-  int nelements;
+  int64_t nelements;
 } cellarray;
