+++ xi_theory/utils/gridlink.h
@@ -3,20 +3,20 @@
 #include "function_precision.h"
 #include "cellarray.h"
 
-cellarray * gridlink(const int np,
-										 const DOUBLE *x,const DOUBLE *y,const DOUBLE *z,
-										 const DOUBLE xmin, const DOUBLE xmax,
-										 const DOUBLE ymin, const DOUBLE ymax,
-										 const DOUBLE zmin, const DOUBLE zmax,
-										 const DOUBLE max_x_size,
-										 const DOUBLE max_y_size,
-										 const DOUBLE max_z_size,
-										 const int xbin_refine_factor,
-										 const int ybin_refine_factor,
-										 const int zbin_refine_factor,
-										 int *nlattice_x,
-										 int *nlattice_y,
-										 int *nlattice_z) __attribute__((warn_unused_result)); 
+cellarray * gridlink(const int64_t np,
+					 const DOUBLE *x,const DOUBLE *y,const DOUBLE *z,
+					 const DOUBLE xmin, const DOUBLE xmax,
+					 const DOUBLE ymin, const DOUBLE ymax,
+					 const DOUBLE zmin, const DOUBLE zmax,
+					 const DOUBLE max_x_size,
+					 const DOUBLE max_y_size,
+					 const DOUBLE max_z_size,
+					 const int xbin_refine_factor,
+					 const int ybin_refine_factor,
+					 const int zbin_refine_factor,
+					 int *nlattice_x,
+					 int *nlattice_y,
+					 int *nlattice_z) __attribute__((warn_unused_result)); 
 
-void get_max_min(const int ND1, const DOUBLE * restrict X1, const DOUBLE * restrict Y1, const DOUBLE * restrict Z1,
-								 DOUBLE *min_x, DOUBLE *min_y, DOUBLE *min_z, DOUBLE *max_x, DOUBLE *max_y, DOUBLE *max_z);
+void get_max_min(const int64_t ND1, const DOUBLE * restrict X1, const DOUBLE * restrict Y1, const DOUBLE * restrict Z1,
+				 DOUBLE *min_x, DOUBLE *min_y, DOUBLE *min_z, DOUBLE *max_x, DOUBLE *max_y, DOUBLE *max_z);
