+++ xi_theory/xi_rp_pi/countpairs_rp_pi.c
@@ -37,14 +37,14 @@ void free_results_rp_pi(results_countpairs_rp_pi **results)
 }
 
 
-results_countpairs_rp_pi * countpairs_rp_pi(const int ND1, const DOUBLE *X1, const DOUBLE *Y1, const DOUBLE *Z1,
-																						const int ND2, const DOUBLE *X2, const DOUBLE *Y2, const DOUBLE *Z2,
+results_countpairs_rp_pi * countpairs_rp_pi(const int64_t ND1, const DOUBLE *X1, const DOUBLE *Y1, const DOUBLE *Z1,
+											const int64_t ND2, const DOUBLE *X2, const DOUBLE *Y2, const DOUBLE *Z2,
 #ifdef USE_OMP
-																						const int numthreads,
+											const int numthreads,
 #endif
-																						const int autocorr,
-																						const char *binfile,
-																						const double pimax)
+											const int autocorr,
+											const char *binfile,
+											const double pimax)
 {
   int bin_refine_factor=1;
 	int zbin_refine_factor=2;
@@ -84,9 +84,9 @@ results_countpairs_rp_pi * countpairs_rp_pi(const int ND1, const DOUBLE *X1, con
 	get_max_min(ND1, X1, Y1, Z1, &xmin, &ymin, &zmin, &xmax, &ymax, &zmax);
 
 	if(autocorr==0) {
-		fprintf(stderr,"ND1 = %8d [xmin,ymin,zmin] = [%lf,%lf,%lf], [xmax,ymax,zmax] = [%lf,%lf,%lf]\n",ND1,xmin,ymin,zmin,xmax,ymax,zmax);
+		fprintf(stderr,"ND1 = %12"PRId64" [xmin,ymin,zmin] = [%lf,%lf,%lf], [xmax,ymax,zmax] = [%lf,%lf,%lf]\n",ND1,xmin,ymin,zmin,xmax,ymax,zmax);
 		get_max_min(ND2, X2, Y2, Z2, &xmin, &ymin, &zmin, &xmax, &ymax, &zmax);
-		fprintf(stderr,"ND2 = %8d [xmin,ymin,zmin] = [%lf,%lf,%lf], [xmax,ymax,zmax] = [%lf,%lf,%lf]\n",ND2,xmin,ymin,zmin,xmax,ymax,zmax);
+		fprintf(stderr,"ND2 = %12"PRId64" [xmin,ymin,zmin] = [%lf,%lf,%lf], [xmax,ymax,zmax] = [%lf,%lf,%lf]\n",ND2,xmin,ymin,zmin,xmax,ymax,zmax);
 	}
 	fprintf(stderr,"Running with [xmin,xmax] = %lf,%lf\n",xmin,xmax);
 	fprintf(stderr,"Running with [ymin,ymax] = %lf,%lf\n",ymin,ymax);
@@ -242,7 +242,7 @@ results_countpairs_rp_pi * countpairs_rp_pi(const int ND1, const DOUBLE *X1, con
 						const DOUBLE *y2 = second->y;
 						const DOUBLE *z2 = second->z;
 						
-						for(int i=0;i<first->nelements;i++){
+						for(int64_t i=0;i<first->nelements;i++){
 							DOUBLE x1pos = x1[i];
 							DOUBLE y1pos = y1[i];
 							DOUBLE z1pos = z1[i];
@@ -253,7 +253,7 @@ results_countpairs_rp_pi * countpairs_rp_pi(const int ND1, const DOUBLE *X1, con
 #endif
 							
 #ifndef USE_AVX	//Beginning of NO AVX section
-							for(int j=0;j<second->nelements;j++) {
+							for(int64_t j=0;j<second->nelements;j++) {
 								const DOUBLE dx = x2[j]-x1pos;
 								const DOUBLE dy = y2[j]-y1pos;
 								const DOUBLE dz = FABS(z2[j]-z1pos);
@@ -300,7 +300,7 @@ results_countpairs_rp_pi * countpairs_rp_pi(const int ND1, const DOUBLE *X1, con
 							const AVX_FLOATS m_y1pos = AVX_SET_FLOAT(y1pos);
 							const AVX_FLOATS m_z1pos = AVX_SET_FLOAT(z1pos);
 
-							int j;
+							int64_t j;
 							for(j=0;j<=(second->nelements-NVEC);j+=NVEC) {
 								const AVX_FLOATS x2pos = AVX_LOAD_FLOATS_UNALIGNED(&x2[j]);
 								const AVX_FLOATS y2pos = AVX_LOAD_FLOATS_UNALIGNED(&y2[j]);
